feat(core+ios): session-builder group management — reps, per-entry co… (#1065)
* feat(core+ios): session-builder group management — reps, per-entry config, sort/remove

Closes the gaps found auditing design/Linked Exercises.dc.html against the
shipped native session builder:

- core: new SessionEvent::AddExerciseToBlock{group_id,item_id} (TDD'd) lets an
  exercise be added into an already-present block, inserted before the anchor
  piece; idempotent by item id like AddToSetlist (#939). FFI bincode
  round-trip coverage added for this event plus the previously-unwired
  SetRepTarget / SetEntryIntention / SetEntryDuration (#846 class).
- ios: SessionBuilderScreen nested rows gain up/down reorder, remove, and a
  settings icon in Edit mode (mirrors LibraryDetailScreen's existing
  edit-row pattern); "Ungroup all" wired to the header.
- ios: new EntrySettingsSheet — per-entry Aim / Track reps / Planned duration,
  the first UI call sites for SetRepTarget/SetEntryIntention/SetEntryDuration.
- ios: new AddRelatedExerciseSheet — single-select picker for adding one more
  exercise into an existing group's block, scoped to unadded exercises.
- ios: incidental fix in KeyPicker.swift — ambiguous CGFloat/Double '+'
  operator that blocked all iOS builds under Xcode 26.3, unrelated to this
  feature but required to build/test locally.

Verified: cargo test (workspace, 460+ core tests green) and cargo clippy
-D warnings clean; swift format clean; full app build + simulator launch;
targeted xcodebuild test run confirms all 4 new snapshot tests pass and no
pre-existing SessionBuilder/AddToSessionSheet test regressed (the only
failures there — including one on a file this diff never touches — carry the
same low-precision signature, consistent with this sandbox's iOS 26.3
simulator vs. the pinned iOS 26.5 snapshot references, not a code regression).

Known follow-up before merge: testSessionBuilderGrouped.1.png needs a real
re-record on the pinned iOS 26.5 toolchain — the new unconditional "Add a
related exercise" footer row on every expanded group legitimately changed
that screen's layout, and this sandbox has no iOS 26.5 runtime available
(Xcode 26.3 here only resolves to 26.3 via `xcodebuild -downloadPlatform`).
Re-record + `just ios-snapshots-optimize`, then run the self-review
(superpowers:code-reviewer / /code-review) and open the PR via `ship`.


* fix(ios): re-record session-builder snapshot references

The new reps/per-entry-config UI in SessionBuilderScreen changed the
grouped layout enough that testSessionBuilderGrouped's existing
reference no longer matched; the 4 new snapshots (AddRelatedExerciseSheet,
EntrySettingsSheet empty/populated, SessionBuilderGroupedEditing) were
apparently recorded on a host that didn't match CI's pinned iOS 26.5
sim, so they failed on first CI run too. Re-recorded all 5 on a
worktree-scoped iOS 26.5 sim (just ios-test) and optimized with
just ios-snapshots-optimize.


---------

+++ ios/Intrada/Views/Components/KeyPicker.swift
@@ -259,7 +259,8 @@ struct KeyPicker: View {
   /// Point at `radius` on the wheel for spoke `ring` (C at top, clockwise).
   private func point(radius: CGFloat, ring: Int) -> CGPoint {
     let radians = (270.0 + 30.0 * Double(ring)) * .pi / 180
-    return CGPoint(x: 150 + radius * cos(radians), y: 150 + radius * sin(radians))
+    return CGPoint(
+      x: 150 + radius * CGFloat(cos(radians)), y: 150 + radius * CGFloat(sin(radians)))
   }
 
   private func tap(ring: Int, mode: Modality) {
